database bugs are fixed

+++ controllers/cache.js
@@ -2,11 +2,20 @@ import NodeCache from "node-cache";
 import CacheData from '../models/cacheData.js';
 import Randomstring from "randomstring";
 
-const newCache = new NodeCache({maxKeys:process.env.MKEYS,stdTTL:process.env.STDTTL,checkperiod:process.env.CHECKPERIOD});
+const newCache = new NodeCache({maxKeys:process.env.MKEYS,stdTTL:process.env.STDTTL});
 
+// whem ttl expired this callback called
+newCache.on("expired",function(key,value){
+    let resObj = { params:{key:key}};
+    const resp = getCachedDataByKey(resObj,res);
+    return resp;
+});
+// Check the maximum limit of cache
 const checkMaxLimit = numberOfKeys => process.env.MAXIMUMKEYS == numberOfKeys ? true : false;
 
 var forward = false;
+
+//Read/create a key
 
 export const getCachedDataByKey = async(req,res) =>{
     
@@ -14,10 +23,19 @@ export const getCachedDataByKey = async(req,res) =>{
 
     const catchData = newCache.get( key )
     if(catchData == undefined ){
+
         console.log('Cache Miss');
+
         const currentKeysCount = newCache.keys().length;
+
         if(checkMaxLimit(currentKeysCount)){
-            newCache.del(newCache.keys()[0]);
+            const oldKey = newCache.keys()[0];
+            try{
+                await CacheData.deleteOne({key:oldKey})
+            }catch(err){
+                return res.json({message:err.message})
+            }
+            newCache.del(oldKey);
         }
         const randomStr = Randomstring.generate();
 
@@ -27,7 +45,7 @@ export const getCachedDataByKey = async(req,res) =>{
         try{
             await newData.save();
         }catch(err){
-            res.status(404).json({message:err.message});
+            return res.status(404).json({message:err.message});
         }
 
         if(forward){
@@ -44,6 +62,8 @@ export const getCachedDataByKey = async(req,res) =>{
     }
 }
 
+// Get All the keys
+
 export const getAllKeys = (req,res) =>{
     try{
         const keys = newCache.keys();
@@ -53,60 +73,62 @@ export const getAllKeys = (req,res) =>{
         console.log(err);
     }
 }
+// create new data for existing key
 
 export const createData = async(req,res) =>{
     const { key,data } = req.body;
     let success;
-    try{
-        forward = false;
-        if(!(newCache.get(key))) {
-            let obj ={ params:{key,data} }
-            forward = true;
-            const randomStr  = getCachedDataByKey(obj,res); 
-            return;
-        }
-        else{
-            success = newCache.set(key,data);
 
-            const newData = CacheData({key:key,data:data})
-            try{
-                await newData.save();
-            }catch(err){
-                res.status(404).json({message:err.message});
-            }
+    forward = false;
+    if(!(newCache.get(key))) {
+        let obj ={ params:{key,data} }
+        forward = true;
+        const randomStr  = getCachedDataByKey(obj,res); 
+        return;
+    }
+    else{
+        success = newCache.set(key,data);
+        try{
+            await CacheData.findOneAndUpdate({key:key},{$set:{data:data}});
+        }catch(err){
+            return res.status(404).json({message:err.message});
         }
-        if(success) return res.json({'message':'Data created'})
     }
-    catch(err){
-        console.log(err);
-    }
+    if(success) return res.json({'message':'Data created'})
+    
 }
-
-/*export const updateDataByKey = async(req,res) =>{
+// update data for a parameter key
+export const updateDataByKey = async(req,res) =>{
 
     const { key } = req.params;
     const { data } = req.body;
 
-    try{
-        const success = newCache.set(key,data);
+    if(!newCache.get(key)) return res.json({'message':'key doesnot exist'})
 
+    try{
+        await CacheData.findByIdAndUpdate({key:key},{$set:{data:data}});
+        const success = newCache.set(key,data);
         if(success) return res.json({'message':'updated'});
     }
     catch(err){
-        console.log(err);
+        return res.json({message:err.message});
     }
-}*/
+}
+
+//Remove a single key,value
 
 export const removeKey = async(req,res) =>{
 
     const { key } = req.params;
     try{
+        if(!newCache.get(key)) return res.json({'message':'key doesnot exist'})
+        
         const result = newCache.del( key );
 
         try{
             await CacheData.findOneAndDelete({key:key});
         }catch(err){
-            res.json({message:err.message});
+            return res.json({message:err.message});
         }
 
         if(result) return res.json({"message":"key removed"});
@@ -116,7 +138,10 @@ export const removeKey = async(req,res) =>{
     }
 }
 
+//Remove all the key value data
+
 export const removeAllKeys = async(req,res) =>{
+    if(newCache.keys().length <= 0) return res.json([])
     try{
         await CacheData.deleteMany({});
         newCache.flushAll();
